Fehlender Text übersetzt und hinzugefügt
"The smallest crack in a hull, the slightest separation between boards offers them entrance. It is not unknown for them to slaughter an entire crew from hold up to deck by infiltrating the ship from underneath.

Fortunately they do not seem to have spread too far outside of the archipelago, but there may come a day when Polpovir attacks become commonplace."