Abilitiestext angepasst
Creates a field of holy energy around the priest, dealing Burn damage and applying Shaken to any vessels in the vicinity while regenerating a modest amount of Health for allies. The power of a priest's Holy Radiance is modified by how well their Reputations align with the preferred behaviors of their faith.

Paladins can command an unconscious ally to get back up with a large instant spike in Health, while also granting the Strong Inspiration. However, the ally will lose half of the regained Health after a short duration.